+++ Modules/Intent.Modules.Application.Contracts/OperationExtensions.cs
@@ -1,19 +1,14 @@
-﻿using System.Linq;
-using Intent.Modelers.Services.Api;
+﻿using Intent.Modelers.Services.Api;
 using Intent.Modules.Application.Contracts.Api;
-using Intent.Modules.Common;
 
 namespace Intent.Modules.Application.Contracts
 {
 
     public static class OperationExtensions
     {
-        // Should come from APi generation
         public static bool IsAsync(this OperationModel operation)
         {
             return !operation.HasSynchronous();
         }
-
-        
     }
 }